requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
PyPDF2==3.0.1
python-docx==1.1.2
google-genai>=1.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
